ARM: tegra: import latest Jetson TK1 pinmux
syseng has revamped the Jetson TK1 pinmux spreadsheet, basing the content
completely on correct configuration for the board/schematic, rather than
the previous version which was based on the bare minimum changes relative
to another reference board.

The new spreadsheet sets TRISTATE for any input-only pins. This only works
correctly if the global CLAMP bit is not set, so don't set it. Apparently
syseng have changed their mind since the previous advice that this needed
to be set:-/ In theory, we should adjust the Venice2 port to also not set
CLAMP, but I don't have time to validate that change at present.

FIXME: describe which tegra-pinmux-scripts version this import came from
FIXME: describe which pinmux XLSM file this came from

Signed-off-by: Stephen Warren <swarren@nvidia.com>